requirements.txt

blinker==1.9.0                                                                                                          
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
easy-scpi==0.1.6
Flask==3.0.3
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
nest-asyncio==1.6.0
numpy==2.2.0
packaging==24.2
pandas==2.2.3
plotly==5.24.1
pyserial==3.5
python-dateutil==2.9.0.post0
pytz==2024.2
PyVISA==1.14.1
requests==2.32.3
retrying==1.3.4
setuptools==75.6.0
six==1.17.0
tenacity==9.0.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.0.6
zipp==3.21.0